Modularized askOptions
