+++ cspell.json
@@ -112,6 +112,7 @@
     "vitepress",
     "vtsls",
     "webfinger",
+    "webframeworks",
     "webp"
   ]
 }
