+++ packages/cli/src/init/ask/pm.ts
@@ -1,0 +1,51 @@
+import { select } from "@inquirer/prompts";
+import { message } from "@optique/core/message";
+import { print } from "@optique/run";
+import * as colors from "@std/fmt/colors";
+import { PACKAGE_MANAGER } from "../const.ts";
+import { getInstallUrl, getLabel, isPackageManagerAvailable } from "../lib.ts";
+import type { PackageManager, WebFramework } from "../types.ts";
+import webFrameworks from "../webframeworks.ts";
+
+const fillPackageManager: //
+  <T extends { packageManager?: PackageManager; webFramework: WebFramework }> //
+  (options: T) => //
+  Promise<Omit<T, "packageManager"> & { packageManager: PackageManager }> = //
+  async ({ packageManager, ...options }) => {
+    const pm = packageManager ?? await askPackageManager(options.webFramework);
+    if (await isPackageManagerAvailable(pm)) {
+      return ({ ...options, packageManager: pm });
+    }
+    noticeInstallUrl(pm);
+    return await fillPackageManager(options) as //
+    typeof options & { packageManager: PackageManager };
+  };
+
+export default fillPackageManager;
+
+const askPackageManager = (wf: WebFramework) =>
+  select<PackageManager>({
+    message: "Choose the package manager to use",
+    choices: PACKAGE_MANAGER.map(choicePackageManager(wf)),
+  });
+
+const choicePackageManager = (wf: WebFramework) => (value: PackageManager) => ({
+  name: isWfSupportsPm(wf, value)
+    ? value
+    : `${value} (not supported with ${webFrameworks[wf].label})`,
+  value,
+  disabled: !isWfSupportsPm(wf, value),
+});
+
+const isWfSupportsPm = (
+  wf: WebFramework,
+  pm: PackageManager,
+) => webFrameworks[wf].packageManagers.includes(pm);
+
+const noticeInstallUrl = (pm: PackageManager) => {
+  const label = colors.bold(getLabel(pm));
+  const url = colors.underline(colors.blue(getInstallUrl(pm)));
+  print(message`  Package manager "${label}" is not installed.`);
+  print(message`  You can install it from following link: ${url}`);
+  print(message`  or choose another package manager:`);
+};
